+++ HyperVBackup/Public/Backup-VM.ps1
@@ -1,35 +1,39 @@
 ﻿function Backup-VM {
     <#
     .SYNOPSIS
-    Perform a backup of Hyper-V Virtual Machines.
+    Performs a backup of one or more Hyper-V Virtual Machines.
 
     .DESCRIPTION
-    Perform a backup of Hyper-V Virtual Machines. This module exports Virtual Machines and compress them using 7zip.
+    Performs a backup of one or more Hyper-V Virtual Machines. This module exports Virtual Machines and compress them using 7zip.
 
     .PARAMETER Name
     Name of the Virtual Machine. You can get a list of your Hyper-V machines using Get-VM.
 
-    .PARAMETER BackupDestination
+    .PARAMETER Destination
     Specify the location of the backup.
+
+    .PARAMETER CompressionLevel
+    Set the compression level of the archive.
+
+    .PARAMETER CompressionMethod
+    Set the compresseion method of the archive.
 
     .PARAMETER Compress
     Specify if you want to compress the backup into a .7z file.
 
-    .PARAMETER Retention
-    Configure how long you want to keep the backups.
-
-    .PARAMETER Force
-    The Force parameter allows the user to skip the "Should Continue" box.
-
     .EXAMPLE
-    Backup-VM -Name "GAME-01", "VPN-01" -BackupDestination 'D:\Backup\Hyper-V' -Retention 30
+    Backup-VM -Name "GAME-01", "VPN-01" -Destination 'D:\Backup\Hyper-V'
+
+    Backups two VMs (GAME-01 and VPN-01) to the destination 'D:\Backup\Hyper-V'.
 
     .EXAMPLE
     Get-VM | Backup-VM -BackupDestination 'D:\Backup\Hyper-V'
 
-    Backups all VMs on the Hyper-V host and store them for 30 days.
+    Backups all VMs on the Hyper-V host to 'D:\Backup\Hyper-V'.
 
     .NOTES
+    Author: Frederik Hjorslev Poulsen
+
     Cmdlet must be executed from an elevated prompt.
 
     .LINK
@@ -44,23 +48,28 @@
         [Parameter(Mandatory = $true,
             ValueFromPipeline = $true,
             ValueFromPipelineByPropertyName = $true,
-            HelpMessage = 'Enter one or more Hyper-V virtual machine names seperated by commas.'
+            HelpMessage = 'Enter one or more Hyper-V virtual machine names seperated by commas.',
+            Position = 0
         )]
         [string[]]$Name,
 
         [Parameter(Mandatory = $true,
-            HelpMessage = 'Enter the backup destination.'
+            HelpMessage = 'Enter the backup destination.',
+            Position = 1
         )]
-        [string]$BackupDestination,
+        [Alias('BackupDestination')]
+        [string]$Destination,
 
         [Parameter(Mandatory = $false)]
-        [switch]$Compress,
+        [ValidateSet('None', 'Fast', 'Low', 'Normal', 'High', 'Ultra')]
+        [string]$CompressionLevel,
 
         [Parameter(Mandatory = $false)]
-        [int]$Retention = 30,
+        [ValidateSet('Copy', 'Deflate', 'Deflate64', 'BZip2', 'Lzma', 'Lzma2', 'Ppmd', 'Default')]
+        [string]$CompressionMethod = 'Default',
 
         [Parameter(Mandatory = $false)]
-        [switch]$Force
+        [switch]$Compress = $true
     )
 
     process {
@@ -70,32 +79,43 @@ Do you want to backup the following Hyper-V machine:
 "@
         if ($Force -or $PSCmdlet.ShouldContinue($BackupSCMessage, 'Backup Virtual Machine?')) {
             Write-Verbose -Message "Virtual machine to backup: $($Name)."
-            Write-Verbose -Message "Backup destination: $($BackupDestination)."
-            Write-Verbose -Message "Days to keep backup: $($Retention)."
+            Write-Verbose -Message "Backup destination: $($Destination)."
             $TimeStamp = Get-Date -Format "yyyy-MM-dd-HHmm"
 
             # Exporting the VM.
-            Export-VM -Name $Name -Path "$($BackupDestination)\$($Name)-$($TimeStamp)"
+            Export-VM -Name $Name -Path "$($Destination)\$($Name)-$($TimeStamp)"
 
             if ($Compress) {
+                # Splat with parameters
+                $CompressParameters = @{
+                    'Path'            = "$($Destination)\$($Name)-$($TimeStamp)";
+                    'ArchiveFileName' = "$($Destination)\$($Name)-$($TimeStamp).7z"
+                }
+
+                # Add CompressionLevel to the splat if variable is populated.
+                if ($null -ne $CompressionLevel) {
+                    Write-Verbose -Message "CompressionLevel set to $($CompressionLevel)."
+
+                    $CompressParameters += @{
+                        'CompressionLevel' = $CompressionLevel
+                    }
+                }
+
+                # Add CompressionMethod to the splat if variable is populated.
+                if ($null -ne $CompressionMethod) {
+                    Write-Verbose -Message "CompressionMethod set to $($CompressionMethod)."
+                    $CompressParameters += @{
+                        'CompressionMethod' = $CompressionMethod
+                    }
+                }
+
                 # Compress the exported VM using 7Zip4PowerShell if the parameter is set.
-                Compress-7Zip -Path "$($BackupDestination)\$($Name)-$($TimeStamp)" -ArchiveFileName "$($BackupDestination)\$($Name)-$($TimeStamp).7z"
+                Write-Verbose -Message 'Backup will be compressed to a .7z archive.'
+                Compress-7Zip @CompressParameters
 
                 # Remove the exported folder as it's now compressed.
-                Remove-Item -Path "$($BackupDestination)\$($Name)-$($TimeStamp)" -Force -Recurse
+                Remove-Item -Path "$($Destination)\$($Name)-$($TimeStamp)" -Force -Recurse
             }
         } # ShouldContinue
     } # Process
-
-    end {
-        $OldBackups = Get-ChildItem -Path $BackupDestination -Recurse -Force | Where-Object -FilterScript { -not $_.PSIsContainer -and (((Get-Date)-$_.LastWriteTime) -gt (New-TimeSpan -Day $Retention)) }
-        $RemoveOldSCMessage = @"
-Do you want to remove any backups older than $($Retention) days?
-"@
-        if ($null -ne $OldBackups) {
-            if ($Force -or $PSCmdlet.ShouldContinue($RemoveOldSCMessage, 'Remove old backups?')) {
-                $OldBackups | Remove-Item -Force
-            }
-        }
-    } # End
 } # Cmdlet